requirements.txt
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.3.1
boto3==1.34.137
botocore==1.34.137
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cohere==5.5.8
cryptography==42.0.8
exceptiongroup==1.2.1
fastavro==1.9.4
filelock==3.15.4
fsspec==2024.6.1
grpcio==1.64.1
grpcio-health-checking==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.7
jmespath==1.0.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
protobuf==5.27.2
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
types-requests==2.32.0.20240622
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.28.3
weaviate-client==4.6.5
